% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_rdepsignore.R
\name{add_to_rdepsignore}
\alias{add_to_rdepsignore}
\title{Add to the .rdepsignore file}
\usage{
add_to_rdepsignore(x)
}
\arguments{
\item{x}{A \code{character} vector. One or several file/folder names to be added
to the \code{.rdepsignore}. This argument is mandatory.}
}
\value{
No return value.
}
\description{
This function adds files/folders to the \code{.rdepsignore} file. The
\code{.rdepsignore} file is used by \code{rdeps} to ignore files or folders when
detecting dependencies. This plain text file contains a list of files/folders
to be ignored.

This file will be created if it does not exist.
}
\examples{
\dontrun{
add_to_rdepsignore("analyses/misc")
add_to_rdepsignore("analyses/old_scripts.R")
}
}
