% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rdeps.R
\name{get_rdeps}
\alias{get_rdeps}
\title{get_rdeps}
\usage{
get_rdeps(pkg, fmt)
}
\arguments{
\item{pkg}{Name of package for which dependencies are to be obtained}

\item{fmt}{If not given, dependencies are returned as simple text list.
Otherwise one of:
\itemize{
\item \code{url} to open all URLs of dependencies where these exist.
Dependencies without associated URLs will also be listed.
\item \code{tar} to download \code{.tar.gz} files of all dependencies.
}}
}
\value{
Character vector listing all dependencies
}
\description{
Automatically obtain source code for all dependencies of a given R package,
either by opening github sites in browser, or downloading .tar.gz files.
}

