% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeps-package.R
\docType{package}
\name{rdeps-package}
\alias{rdeps}
\alias{rdeps-package}
\title{rdeps: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/rdeps}
  \item Report bugs at \url{https://github.com/frbcesab/rdeps/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
