% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeps.R
\docType{package}
\name{rdeps}
\alias{rdeps}
\alias{rdeps-package}
\title{rdeps}
\description{
Automatically obtain source code for all dependencies of a given R package,
either by opening github sites in browser, or downloading .tar.gz files.
}
\section{Functions}{

\tabular{ll}{
\code{get_rdeps}\tab Obtain source code
}
}

