% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeps-package.R
\docType{package}
\name{rdeps-package}
\alias{rdeps}
\alias{rdeps-package}
\title{rdeps: Identify External Packages Used in a Project}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Screens all '.R', '.Rmd', and '.qmd' files to extract the name of packages used in a project. This package detects packages called with 'library(foo)', 'require(foo)', and 'foo::bar()' and adds these dependencies in the DESCRIPTION file in the sections Depends, Imports, and Suggests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/rdeps}
  \item Report bugs at \url{https://github.com/frbcesab/rdeps/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
